+++ TLSAcceptor.go
@@ -4,16 +4,21 @@ import (
 	"crypto/tls"
 	"net"
 	"log"
+	"fmt"
 )
 
 type TLSAcceptor struct {
 	listener	net.Listener
-	conns		chan net.Conn
+	conns		chan *tls.Conn
+	config		*tls.Config
+	port		int
 }
 
-func NewTLSAcceptor(portStr string) (*TLSAcceptor,error) {
+func NewTLSAcceptor(port int) (*TLSAcceptor,error) {
 	accp := new(TLSAcceptor)
-	accp.conns = make(chan net.Conn,100)
+	accp.port = port
+	portStr := fmt.Sprintf("%v",port)
+	accp.conns = make(chan *tls.Conn,100)
 	certbs,keybs := GenerateTLSKeys()
 	cert,err := tls.X509KeyPair(certbs,keybs)
 	if err!=nil {
@@ -21,6 +26,9 @@ func NewTLSAcceptor(portStr string) (*TLSAcceptor,error) {
 	}
 	config := new(tls.Config)
 	config.Certificates = []tls.Certificate{cert}
+	config.ClientAuth = tls.RequestClientCert
+	config.InsecureSkipVerify = true
+	accp.config = config
 	accp.listener,err = tls.Listen("tcp",":"+portStr,config)
 	if err!=nil {
 		return nil,err
@@ -32,12 +40,24 @@ func NewTLSAcceptor(portStr string) (*TLSAcceptor,error) {
 				log.Print("ERROR::TLSAcceptor:: ",err)
 				continue
 			}
-			accp.conns <- conn
+			accp.conns <- conn.(*tls.Conn)
 		}
 	}()
 	return accp,nil
 }
 
-func (accp *TLSAcceptor) GetConn() net.Conn {
-	return <-accp.conns
+func (accp *TLSAcceptor) GetConn() *BroConn {
+	conn :=  <-accp.conns
+	log.Print("accepted new conn. Build BroConn...")
+	broConn := NewBroConn(nil,accp.port,accp.config)
+	err := broConn.SetConn(conn)
+	if err!=nil {
+		log.Print(err)
+		return nil
+	}
+	return broConn
+}
+
+func (accp *TLSAcceptor) GetConfig() *tls.Config {
+	return accp.config
 }