+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+04. Mar. 2013
+- Added whispering with connection recycling
+- Added /help command
+- Fixed some things ;)
+
 27. Feb. 2013
 - Added TLSAcceptor
 - cleaned up UI Interface